requirements.txt
fastapi
uvicorn
websockets
pydantic
langchain
langchain-openai
langchain-mongodb
python-dotenv